Add files via upload
Added dog tests, to prepare for infobox about dog.
Buttons added, currently functionless except for infobox dog button.

Since the coordinates are always the same no matter where you scroll, it's difficult to get the Dog button on to the actual dog. Testing with constraints to ensure button is only clickable in sight of Dog.